Replace (unsigned)0 with 0u

+++ Ex_2.6/Ex_2.6.cpp
@@ -11,17 +11,17 @@ setbits(target_number, pos_from, bit_amount, source_number)
 // functions
 // ###########################################################################
 
-unsigned int setbits(unsigned int target_number, int pos_from, int bit_amount, unsigned int source_number);
+unsigned long setbits(unsigned long target_number, int pos_from, int bit_amount, unsigned long source_number);
 
 // ###########################################################################
 
 int main()
 {
-	unsigned int target_number = 0b10110101;
+	unsigned long target_number = 0b10110101;
 	int bit_amount = 4;
 	int pos_from = 6;
-	unsigned int source_number = 0b11000111;
-	unsigned int result_number = 0;
+	unsigned long source_number = 0b11000111;
+	unsigned long result_number = 0;
 
 	result_number = setbits(target_number, pos_from, bit_amount, source_number);
 	printf("%d", result_number);
@@ -29,21 +29,21 @@ int main()
 }
 
 
-unsigned int setbits(unsigned int target_number, int pos_from, int bit_amount, unsigned int source_number)
+unsigned long setbits(unsigned long target_number, int pos_from, int bit_amount, unsigned long source_number)
 {
-	unsigned int part_before_pos = 0;
-	unsigned int part_after_pos = 0;
+	unsigned long part_before_pos = 0;
+	unsigned long part_after_pos = 0;
 
 	//Mask rightmost bit_amount bits from sourse_number and move it to the pos_from in target_number
-	source_number = (source_number & ~(~(unsigned)0 << bit_amount)) << (pos_from + 1 - bit_amount);
+	source_number = (source_number & ~(~ 0u << bit_amount)) << (pos_from + 1 - bit_amount);
 	//printf("%d", source_number);
 
 	//Target_number can be represented as: part_before_pos + pos_from + part_after_pos
 	//Mask part_before_pos
-	part_before_pos = target_number & (~(unsigned)0 << (pos_from + 1));
+	part_before_pos = target_number & (~ 0u << (pos_from + 1));
 	
 	//Mask part_after_pos
-	part_after_pos = target_number & ~(~(unsigned)0 << (pos_from + 1 - bit_amount));
+	part_after_pos = target_number & ~(~ 0u << (pos_from + 1 - bit_amount));
 
 	//Set to one in target_number bits that are set to one in part_before_pos, part_after_pos and source_number
 	target_number = part_before_pos | part_after_pos | source_number;
